WebApp/requirements.txt
fastapi
uvicorn
streamlit
Pillow
requests
